fix withdraw, can only withdraw int

+++ src/main/java/rocks/zipcode/atm/bank/Bank.java
@@ -42,7 +42,7 @@ public class Bank {
         return ActionResult.success(account.getAccountData());
     }
 
-    public ActionResult<AccountData> withdraw(AccountData accountData, float amount) {
+    public ActionResult<AccountData> withdraw(AccountData accountData, int amount) {
         Account account = accounts.get(accountData.getId());
         boolean ok = account.withdraw(amount);
